# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClearCredit API",
    "description": "Carbon-credit integrity and double-counting checker. Produces integrity scores and verified integrity attestations; it does not certify emission reductions.",
    "version": "0.1.0"
  },
  "paths": {
    "/claims/preview": {
      "post": {
        "tags": [
          "claims"
        ],
        "summary": "Preview",
        "description": "Validate, check overlaps, gather evidence and score \u2014 without registering anything.",
        "operationId": "preview_claims_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Claim"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "413": {
            "description": "REQUEST_TOO_LARGE: request bodies are limited to 2 MiB before parsing."
          },
          "503": {
            "description": "EVIDENCE_INVALID, CHAIN_UNAVAILABLE or CHAIN_READ_UNAVAILABLE: evidence review or retry of required chain state is necessary."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/claims": {
      "post": {
        "tags": [
          "claims"
        ],
        "summary": "Submit",
        "description": "Register a signed claim. Requires `Idempotency-Key`; a retry with the same key and body\nreturns the original result and never creates a second record or transaction.",
        "operationId": "submit_claims_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Submission"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "413": {
            "description": "REQUEST_TOO_LARGE: request bodies are limited to 2 MiB before parsing."
          },
          "503": {
            "description": "EVIDENCE_INVALID, CHAIN_UNAVAILABLE or CHAIN_READ_UNAVAILABLE: evidence review or retry of required chain state is necessary."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/claims/{ref}": {
      "get": {
        "tags": [
          "claims"
        ],
        "summary": "Get Claim",
        "description": "Claim by project ID or claim hash: canonical hash, overlaps, evidence, score, on-chain state.",
        "operationId": "get_claim_claims__ref__get",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          },
          {
            "name": "minBlock",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Minblock"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "413": {
            "description": "REQUEST_TOO_LARGE: request bodies are limited to 2 MiB before parsing."
          },
          "503": {
            "description": "EVIDENCE_INVALID, CHAIN_UNAVAILABLE or CHAIN_READ_UNAVAILABLE: evidence review or retry of required chain state is necessary."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/claims/{ref}/verify": {
      "get": {
        "tags": [
          "claims"
        ],
        "summary": "Verify",
        "description": "Recompute the hash from the stored claim and compare it with the on-chain value.",
        "operationId": "verify_claims__ref__verify_get",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          },
          {
            "name": "minBlock",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Minblock"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "413": {
            "description": "REQUEST_TOO_LARGE: request bodies are limited to 2 MiB before parsing."
          },
          "503": {
            "description": "EVIDENCE_INVALID, CHAIN_UNAVAILABLE or CHAIN_READ_UNAVAILABLE: evidence review or retry of required chain state is necessary."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/claims/{ref}/history": {
      "get": {
        "tags": [
          "claims"
        ],
        "summary": "Claim History",
        "description": "Project events in an explicit window (at most 10,000 blocks / 1,000 events).",
        "operationId": "claim_history_claims__ref__history_get",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          },
          {
            "name": "fromBlock",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fromblock"
            }
          },
          {
            "name": "toBlock",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Toblock"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "413": {
            "description": "REQUEST_TOO_LARGE: request bodies are limited to 2 MiB before parsing."
          },
          "503": {
            "description": "EVIDENCE_INVALID, CHAIN_UNAVAILABLE or CHAIN_READ_UNAVAILABLE: evidence review or retry of required chain state is necessary."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/overlaps/{ref}": {
      "get": {
        "tags": [
          "claims"
        ],
        "summary": "Overlaps",
        "operationId": "overlaps_overlaps__ref__get",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "413": {
            "description": "REQUEST_TOO_LARGE: request bodies are limited to 2 MiB before parsing."
          },
          "503": {
            "description": "EVIDENCE_INVALID, CHAIN_UNAVAILABLE or CHAIN_READ_UNAVAILABLE: evidence review or retry of required chain state is necessary."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Submission": {
        "properties": {
          "claim": {
            "$ref": "#/components/schemas/Claim"
          },
          "signature": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{130}$",
            "title": "Signature",
            "description": "Developer's EIP-712 signature over the Claim typed data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "claim",
          "signature"
        ],
        "title": "Submission"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}